+++ TelaRela.h
@@ -29,7 +29,7 @@
   class TelaRela {
     public:
       TelaRela(void); // Configura os componentes da tela
-      void atualizaTelaRela(uint8_t minuto, uint8_t segundo, uint8_t miliSegundo, bool proximo);
+      void atualizaTelaRela(uint8_t minuto, uint8_t segundo, uint32_t miliSegundo, bool proximo);
       void zeraRelatorios(void);
     private:
       static uint8_t posicao;
@@ -43,8 +43,8 @@
       bool getProximo(void);
       void setProximo(bool value);
       void proximoTempo(void);
-      void adicionaTempo(uint8_t minuto, uint8_t segundo, uint8_t miliSegundo, uint8_t posicao);
-      bool tempoZero(uint8_t minuto, uint8_t segundo, uint8_t miliSegundo);
+      void adicionaTempo(uint8_t minuto, uint8_t segundo, uint32_t miliSegundo, uint8_t posicao);
+      bool tempoZero(uint8_t minuto, uint8_t segundo, uint32_t miliSegundo);
   };
 
 #endif //TELA_RELA_H
